+++ lib/src/main/java/com/zhang/library/utils/HandlerManager.java
@@ -1,4 +1,4 @@
-package com.zhang.library.utils.utils;
+package com.zhang.library.utils;
 
 import android.os.Handler;
 import android.os.HandlerThread;
