+++ lib/src/main/java/com/zhang/library/utils/TimeUtils.java
@@ -1,4 +1,4 @@
-package com.zhang.library.utils.utils;
+package com.zhang.library.utils;
 
 import java.text.DateFormat;
 import java.text.ParseException;
