+++ lib/src/main/java/com/zhang/library/utils/context/ClipboardUtils.java
@@ -1,4 +1,4 @@
-package com.zhang.library.utils.utils.context;
+package com.zhang.library.utils.context;
 
 import android.content.ClipData;
 import android.content.ClipDescription;
