+++ lib/src/main/java/com/zhang/library/utils/context/GlideUtils.java
@@ -1,4 +1,4 @@
-package com.zhang.library.utils.utils.context;
+package com.zhang.library.utils.context;
 
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
