+++ lib/src/main/java/com/zhang/library/utils/context/ResUtils.java
@@ -1,4 +1,4 @@
-package com.zhang.library.utils.utils.context;
+package com.zhang.library.utils.context;
 
 import android.app.Activity;
 import android.content.Context;
