+++ lib/src/main/java/com/zhang/library/utils/context/StringUtils.java
@@ -1,4 +1,4 @@
-package com.zhang.library.utils.utils.context;
+package com.zhang.library.utils.context;
 
 import android.text.Spannable;
 import android.text.SpannableString;
